Extract Workloadhints tests
As these tests seems to take a lot of the execution time lets extract
them so we could execute them in a different lane.
