+++ Makefile
@@ -210,6 +210,15 @@ pao-functests-update-only:
 	hack/show-cluster-version.sh
 	hack/run-test.sh -t "test/e2e/performanceprofile/functests/0_config test/e2e/performanceprofile/functests/2_performance_update test/e2e/performanceprofile/functests/3_performance_status test/e2e/performanceprofile/functests/7_performance_kubelet_node test/e2e/performanceprofile/functests/8_reboot" -p "-v -r -timeout=5h --fail-fast --flake-attempts=2" -m "Running Functional Tests" -r "--junit-report=report.xml"
 
+.PHONY: pao-functests-performance-workloadhints
+pao-functests-performance-workloadhints: cluster-label-worker-cnf pao-functests-performance-workloadhints-only
+
+.PHONY: pao-functests-performance-workloadhints-only
+pao-functests-performance-workloadhints-only:
+	@echo "Cluster Version"
+	hack/show-cluster-version.sh
+	hack/run-test.sh -t "test/e2e/performanceprofile/functests/0_config test/e2e/performanceprofile/functests/8_performance_workloadhints" -p "-v -r -timeout=5h --fail-fast --flake-attempts=2" -m "Running Functional WorkloadHints Tests" -r "--junit-report=report.xml"
+
 .PHONY: pao-functests-latency-testing
 pao-functests-latency-testing: dist-latency-tests
 	@echo "Cluster Version"
