+++ test/e2e/performanceprofile/functests/8_performance_workloadhints/test_suite_performance_workloadhints_test.go
@@ -1,0 +1,51 @@
+//go:build !unittests
+// +build !unittests
+
+package __performance_workloadhints_test
+
+import (
+	"context"
+	"testing"
+	"time"
+
+	. "github.com/onsi/ginkgo/v2"
+	"github.com/onsi/ginkgo/v2/reporters"
+	. "github.com/onsi/gomega"
+
+	"k8s.io/apimachinery/pkg/api/errors"
+
+	qe_reporters "kubevirt.io/qe-tools/pkg/ginkgo-reporters"
+
+	testutils "github.com/openshift/cluster-node-tuning-operator/test/e2e/performanceprofile/functests/utils"
+	testclient "github.com/openshift/cluster-node-tuning-operator/test/e2e/performanceprofile/functests/utils/client"
+	testlog "github.com/openshift/cluster-node-tuning-operator/test/e2e/performanceprofile/functests/utils/log"
+	"github.com/openshift/cluster-node-tuning-operator/test/e2e/performanceprofile/functests/utils/namespaces"
+)
+
+var _ = BeforeSuite(func() {
+	// create test namespace
+	err := testclient.Client.Create(context.TODO(), namespaces.TestingNamespace)
+	if errors.IsAlreadyExists(err) {
+		testlog.Warning("test namespace already exists, that is unexpected")
+		return
+	}
+	Expect(err).ToNot(HaveOccurred())
+})
+
+var _ = AfterSuite(func() {
+	err := testclient.Client.Delete(context.TODO(), namespaces.TestingNamespace)
+	Expect(err).ToNot(HaveOccurred())
+	err = namespaces.WaitForDeletion(testutils.NamespaceTesting, 5*time.Minute)
+})
+
+func TestPerformanceUpdate(t *testing.T) {
+	RegisterFailHandler(Fail)
+
+	RunSpecs(t, "Performance Addon Operator WorkloadHints")
+}
+
+var _ = ReportAfterSuite("PAO workloadhints suite", func(r Report) {
+	if qe_reporters.Polarion.Run {
+		reporters.ReportViaDeprecatedReporter(&qe_reporters.Polarion, r)
+	}
+})
